+++ components/icon/index.tsx
@@ -1,6 +1,6 @@
 import warning from '../_util/warning';
 
-const Icon = () => {
+const Icon: React.FC<any> = () => {
   warning(false, 'Icon', 'Empty Icon');
   return null;
 };
