+++ components/radio/__tests__/group.test.tsx
@@ -1,47 +1,27 @@
 import React from 'react';
-import Radio, { Button } from '..';
-import focusTest from '../../../tests/shared/focusTest';
-import mountTest from '../../../tests/shared/mountTest';
-import rtlTest from '../../../tests/shared/rtlTest';
-
+import type { RefAttributes } from 'react';
+import type { RadioGroupProps } from '..';
 import { render, fireEvent } from '../../../tests/utils';
-
-describe('Radio Button', () => {
-  focusTest(Button, { refFocus: true });
-  mountTest(Button);
-
-  rtlTest(Button);
-
-  it('should render correctly', () => {
-    const { container } = render(<Button className="customized">Test</Button>);
-    expect(container.firstChild).toMatchSnapshot();
-  });
-
-  it('responses hover events', () => {
-    const onMouseEnter = jest.fn();
-    const onMouseLeave = jest.fn();
-
-    const { container } = render(
-      <Button onMouseEnter={onMouseEnter} onMouseLeave={onMouseLeave} />,
-    );
-
-    fireEvent.mouseEnter(container.querySelector('label'));
-    expect(onMouseEnter).toHaveBeenCalled();
-
-    fireEvent.mouseLeave(container.querySelector('label'));
-    expect(onMouseLeave).toHaveBeenCalled();
-  });
-});
+import Radio from '..';
 
 describe('Radio Group', () => {
-  function createRadioGroup(props) {
+  function createRadioGroup(props?: RadioGroupProps & RefAttributes<HTMLDivElement>) {
     return (
       <Radio.Group {...props}>
-        <Button value="A">A</Button>
-        <Button value="B">B</Button>
-        <Button value="C">C</Button>
+        <Radio value="A">A</Radio>
+        <Radio value="B">B</Radio>
+        <Radio value="C">C</Radio>
       </Radio.Group>
     );
+  }
+
+  function createRadioGroupByOption(props?: RadioGroupProps & RefAttributes<HTMLDivElement>) {
+    const options = [
+      { label: 'A', value: 'A' },
+      { label: 'B', value: 'B' },
+      { label: 'C', value: 'C' },
+    ];
+    return <Radio.Group {...props} options={options} />;
   }
 
   it('responses hover events', () => {
@@ -54,22 +34,30 @@ describe('Radio Group', () => {
       </Radio.Group>,
     );
 
-    fireEvent.mouseEnter(container.querySelectorAll('div')[0]);
+    fireEvent.mouseEnter(container.querySelector('div')!);
     expect(onMouseEnter).toHaveBeenCalled();
 
-    fireEvent.mouseLeave(container.querySelectorAll('div')[0]);
+    fireEvent.mouseLeave(container.querySelector('div')!);
     expect(onMouseLeave).toHaveBeenCalled();
   });
 
   it('fire change events when value changes', () => {
     const onChange = jest.fn();
 
-    const { container, rerender } = render(createRadioGroup({ onChange }));
-
+    const { container, rerender } = render(
+      createRadioGroup({
+        onChange,
+      }),
+    );
     const radios = container.querySelectorAll('input');
 
     // controlled component
-    rerender(createRadioGroup({ onChange, value: 'A' }));
+    rerender(
+      createRadioGroup({
+        onChange,
+        value: 'A',
+      }),
+    );
     fireEvent.click(radios[1]);
     expect(onChange.mock.calls.length).toBe(1);
   });
@@ -78,47 +66,50 @@ describe('Radio Group', () => {
     const onChange = jest.fn();
     const onChangeRadioGroup = jest.fn();
 
-    const { container } = render(
-      <Radio.Group onChange={onChangeRadioGroup}>
-        <Radio value="A" onChange={onChange}>
+    const RadioGroup: React.FC<
+      RadioGroupProps & { onChangeRadioGroup: RadioGroupProps['onChange'] }
+    > = props => (
+      <Radio.Group onChange={props.onChangeRadioGroup}>
+        <Radio value="A" onChange={props.onChange}>
           A
         </Radio>
-        <Radio value="B" onChange={onChange}>
+        <Radio value="B" onChange={props.onChange}>
           B
         </Radio>
-        <Radio value="C" onChange={onChange}>
+        <Radio value="C" onChange={props.onChange}>
           C
         </Radio>
-      </Radio.Group>,
+      </Radio.Group>
+    );
+
+    const { container, rerender } = render(
+      <RadioGroup onChangeRadioGroup={onChangeRadioGroup} onChange={onChange} />,
     );
     const radios = container.querySelectorAll('input');
 
     // controlled component
+    rerender(<RadioGroup value="A" onChangeRadioGroup={onChangeRadioGroup} onChange={onChange} />);
     fireEvent.click(radios[1]);
     expect(onChange.mock.calls.length).toBe(1);
     expect(onChangeRadioGroup.mock.calls.length).toBe(1);
   });
 
-  it('Trigger onChange when both of Button and radioGroup exists', () => {
+  it('Trigger onChange when both of radioButton and radioGroup exists', () => {
     const onChange = jest.fn();
 
-    const { container, rerender } = render(
-      <Radio.Group onChange={onChange}>
-        <Button value="A">A</Button>
-        <Button value="B">B</Button>
-        <Button value="C">C</Button>
-      </Radio.Group>,
+    const RadioGroup: React.FC<RadioGroupProps> = props => (
+      <Radio.Group {...props}>
+        <Radio.Button value="A">A</Radio.Button>
+        <Radio.Button value="B">B</Radio.Button>
+        <Radio.Button value="C">C</Radio.Button>
+      </Radio.Group>
     );
+
+    const { container, rerender } = render(<RadioGroup onChange={onChange} />);
     const radios = container.querySelectorAll('input');
 
     // controlled component
-    rerender(
-      <Radio.Group value="A" onChange={onChange}>
-        <Button value="A">A</Button>
-        <Button value="B">B</Button>
-        <Button value="C">C</Button>
-      </Radio.Group>,
-    );
+    rerender(<RadioGroup value="A" onChange={onChange} />);
     fireEvent.click(radios[1]);
     expect(onChange.mock.calls.length).toBe(1);
   });
@@ -128,7 +119,7 @@ describe('Radio Group', () => {
     const options = [{ label: 'Bamboo', value: 'Bamboo' }];
     const { container } = render(<Radio.Group options={options} onChange={onChange} />);
 
-    fireEvent.click(container.querySelector('input'));
+    fireEvent.click(container.querySelector('input')!);
     expect(onChange).toHaveBeenCalledTimes(1);
   });
 
@@ -143,16 +134,28 @@ describe('Radio Group', () => {
     const radios = container.querySelectorAll('input');
 
     // controlled component
-    rerender(createRadioGroup({ onChange, value: 'A' }));
+    rerender(
+      createRadioGroup({
+        onChange,
+        value: 'A',
+      }),
+    );
     fireEvent.click(radios[0]);
     expect(onChange.mock.calls.length).toBe(0);
+  });
+
+  it('optional should correct render', () => {
+    const { container } = render(createRadioGroupByOption());
+    const radios = container.querySelectorAll('input');
+
+    expect(radios.length).toBe(3);
   });
 
   it('all children should have a name property', () => {
     const GROUP_NAME = 'GROUP_NAME';
     const { container } = render(createRadioGroup({ name: GROUP_NAME }));
 
-    container.querySelectorAll('input[type="radio"]').forEach(el => {
+    container.querySelectorAll<HTMLInputElement>('input[type="radio"]').forEach(el => {
       expect(el.name).toEqual(GROUP_NAME);
     });
   });
@@ -167,16 +170,16 @@ describe('Radio Group', () => {
   });
 
   it('should forward ref', () => {
-    let radioGroupRef;
+    let radioGroupRef: HTMLDivElement;
     const { container } = render(
-      createRadioGroup({
-        ref: ref => {
+      createRadioGroupByOption({
+        ref(ref: HTMLDivElement) {
           radioGroupRef = ref;
         },
       }),
     );
 
-    expect(radioGroupRef).toBe(container.querySelector('.ant-radio-group'));
+    expect(radioGroupRef!).toBe(container.querySelector<HTMLDivElement>('.ant-radio-group'));
   });
 
   it('should support data-* or aria-* props', () => {
@@ -184,10 +187,14 @@ describe('Radio Group', () => {
       createRadioGroup({
         'data-radio-group-id': 'radio-group-id',
         'aria-label': 'radio-group',
-      }),
+      } as RadioGroupProps),
     );
-    expect(container.firstChild.getAttribute('data-radio-group-id')).toBe('radio-group-id');
-    expect(container.firstChild.getAttribute('aria-label')).toBe('radio-group');
+    expect((container.firstChild as HTMLDivElement)?.getAttribute('data-radio-group-id')).toBe(
+      'radio-group-id',
+    );
+    expect((container.firstChild as HTMLDivElement)?.getAttribute('aria-label')).toBe(
+      'radio-group',
+    );
   });
 
   it('Radio type should not be override', () => {
@@ -216,29 +223,33 @@ describe('Radio Group', () => {
 
   describe('value is null or undefined', () => {
     it('use `defaultValue` when `value` is undefined', () => {
+      const options = [{ label: 'Bamboo', value: 'bamboo' }];
       const { container } = render(
-        <Radio.Group defaultValue="bamboo" value={undefined}>
-          <Button value="bamboo">Bamboo</Button>
-        </Radio.Group>,
+        <Radio.Group defaultValue="bamboo" value={undefined} options={options} />,
       );
-      expect(container.querySelectorAll('.ant-radio-button-wrapper-checked').length).toBe(1);
+      expect(container.querySelectorAll('.ant-radio-wrapper-checked').length).toBe(1);
     });
 
     [undefined, null].forEach(newValue => {
       it(`should set value back when value change back to ${newValue}`, () => {
-        const { container, rerender } = render(
-          <Radio.Group value="bamboo">
-            <Button value="bamboo">Bamboo</Button>
-          </Radio.Group>,
-        );
-        expect(container.querySelectorAll('.ant-radio-button-wrapper-checked').length).toBe(1);
-        rerender(
-          <Radio.Group value={newValue}>
-            <Button value="bamboo">Bamboo</Button>
-          </Radio.Group>,
-        );
-        expect(container.querySelectorAll('.ant-radio-button-wrapper-checked').length).toBe(0);
+        const options = [{ label: 'Bamboo', value: 'bamboo' }];
+        const { container, rerender } = render(<Radio.Group value="bamboo" options={options} />);
+        expect(container.querySelectorAll('.ant-radio-wrapper-checked').length).toBe(1);
+        rerender(<Radio.Group value={newValue} options={options} />);
+        expect(container.querySelectorAll('.ant-radio-wrapper-checked').length).toBe(0);
       });
     });
+  });
+
+  it('onBlur & onFocus should work', () => {
+    const handleBlur = jest.fn();
+    const handleFocus = jest.fn();
+    const { container } = render(
+      <Radio.Group options={['1', '2', '3']} onBlur={handleBlur} onFocus={handleFocus} />,
+    );
+    fireEvent.focus(container.firstChild!);
+    expect(handleFocus).toHaveBeenCalledTimes(1);
+    fireEvent.blur(container.firstChild!);
+    expect(handleBlur).toHaveBeenCalledTimes(1);
   });
 });
