+++ components/rate/__tests__/demo.test.ts
