+++ components/rate/__tests__/index.test.ts
