+++ components/slider/__tests__/demo.test.ts
